deployment/requirements.txt
absl-py==0.11.0
astor==0.8.1
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gast==0.2.2
google-api-core==1.24.1
google-auth==1.24.0
google-cloud==0.34.0
google-cloud-core==1.5.0
google-cloud-storage==1.35.0
google-crc32c==1.1.0
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
h5py==3.1.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
networkx==2.5
numpy==1.19.4
opencv-python==4.5.1.48
opt-einsum==3.3.0
Pillow==8.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
requests==2.25.1
rsa==4.6
scikit-image==0.18.1
scipy==1.6.0
six==1.15.0
SQLAlchemy==1.3.22
tensorboard==1.15.0
tensorflow-cpu==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
tifffile==2020.12.8
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.0
